housekeeping: refactor task board screen into separate files
